refactor: CurrentValidWeight 类型从 decimal? 改为 int?
有效投料重量改为整数值，去掉小数位。涉及实体、DTO、Blazor、
Service、Test 共 21 个文件，均为类型声明和字面量适配。

+++ MES.Core/DTOs/Batch/CreateProductionBatchRequest.cs
@@ -65,7 +65,7 @@ public class CreateProductionBatchRequest
 
     public int? CurrentValidQty { get; set; }
 
-    public decimal? CurrentValidWeight { get; set; }
+    public int? CurrentValidWeight { get; set; }
 
     // ========== 工单字段（前端可修改，不传则从工单自动复制） ==========
